make build-fast
Build time exceeds Travis 50 minute limit, so we add a `build-fast`
target to the Makefile and the .travis.yaml which passes the --fast
argument to stack build in the Dockerfile when building under Travis.

With --fast, Travis build time is 45 min 3 sec.

Also stack exec ihaskell -- install so that we don't make an unneeded
copy of the ihaskell binary.